feat: 알림버튼 누르면 알림모달 열고 닫히는 기능 구현
