+++ src/components/organisms/notification-list/NotificationList.tsx
@@ -1,16 +1,16 @@
-import AlertCard from '@/components/molecules/card/AlertCard';
+import NotificationCard from '@/components/molecules/card/NotificationCard';
 import { Notification } from '@/types/notification';
 
 interface Props {
   notificationList: Notification[];
 }
 
-export default function AlertList({ notificationList }: Props) {
+export default function NotificationList({ notificationList }: Props) {
   if (!notificationList) return <p>알림이 없습니다.</p>;
   return (
     <ol className="flex flex-col gap-[8px]">
       {notificationList.map((notification) => (
-        <AlertCard key={notification.id} notification={notification} />
+        <NotificationCard key={notification.id} notification={notification} />
       ))}
     </ol>
   );
